MLK-20201 clk: imx: Fix coverity issue of CID 17326/17327
This patch fixes below coverity issues:

CID 17326/17327: Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)
overflow_before_widen: Potentially overflowing expression div *
parent_rate with type unsigned long (32 bits, unsigned) is evaluated
using 32-bit arithmetic, and then used in a context that expects an
expression of type u64 (64 bits, unsigned).

+++ drivers/clk/imx/clk-pllv3.c
@@ -298,7 +298,7 @@ static long clk_pllv3_av_round_rate(struct clk_hw *hw, unsigned long rate,
 		mfd = parent_rate;
 
 	div = rate / parent_rate;
-	temp64 = (u64) (rate - div * parent_rate);
+	temp64 = rate - div * parent_rate;
 	temp64 *= mfd;
 	do_div(temp64, parent_rate);
 	mfn = temp64;
@@ -328,7 +328,7 @@ static int clk_pllv3_av_set_rate(struct clk_hw *hw, unsigned long rate,
 		mfd = parent_rate;
 
 	div = rate / parent_rate;
-	temp64 = (u64) (rate - div * parent_rate);
+	temp64 = rate - div * parent_rate;
 	temp64 *= mfd;
 	do_div(temp64, parent_rate);
 	mfn = temp64;
